updated fish functions, neofetch config, pwsh installer, conky

+++ Windows/install.ps1
@@ -32,21 +32,32 @@ $curDir = Split-Path -Parent $MyInvocation.MyCommand.Definition
 
 # Pass current script to admin console if current console not elevated
 If (-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] "Administrator")) {   
-  Start-Process -WindowStyle hidden powershell -Verb runAs -ArgumentList $curDir\setup.ps1
-  alertUser
+  #Start-Process -WindowStyle hidden powershell -Verb runAs -ArgumentList $curDir\setup.ps1
+  #alertUser
+  Write-Error -Message "Installer must be run as admin." -Category AuthenticationError
 }
 # Otherwise, continue as-is
 else {
   # Copy files into appropriate windows directory
-  # TODO: add user option for what to install like the linux installer does
+
+  $pfunc = Read-Host -Prompt 'Install pwsh profile/functions? (y/N)'
+  $ahk = Read-Host -Prompt 'Install ahk scripts? (y/N)'
+  $wterm = Read-Host -Prompt 'Install Terminal config? (y/N)'
   
   Set-Location $curDir
 
-  Copy-Item -Force ".\profile.ps1" -Destination "$env:windir\system32\WindowsPowerShell\v1.0\"
+  if ($pfunc -eq "y" -or "Y") {
+    Copy-Item -Force ".\profile.ps1" -Destination "$env:windir\system32\WindowsPowerShell\v1.0\"
 
-  Copy-Item -Force .\functions -Destination $env:windir\system32\WindowsPowerShell\v1.0\ -Recurse
+    Copy-Item -Force .\functions -Destination $env:windir\system32\WindowsPowerShell\v1.0\ -Recurse
+  }
+
+  if ($ahk -eq "y" -or "Y") {
+    Copy-Item -Force '..\..\AutoHotkey\Keyboard Shortcuts\win_tweaks.exe' -Destination "$env:APPDATA\Microsoft\Windows\Start Menu\Programs\Startup"
+  }
   
-  Copy-Item -Force '..\..\AutoHotkey\Keyboard Shortcuts\win_tweaks.exe' -Destination "$env:APPDATA\Microsoft\Windows\Start Menu\Programs\Startup"
 
-  Copy-Item -Force '.\profiles.json' -Destination "$env:HOMEPATH\AppData\Local\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\RoamingState"
+  if ($wterm -eq "y" -or "Y") {
+    Copy-Item -Force '.\profiles.json' -Destination "$env:HOMEPATH\AppData\Local\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\RoamingState"
+  }
 }
